Add JointModel
